+++ Mapping/AttributeMetadata.php
@@ -15,6 +15,8 @@ use Symfony\Component\PropertyAccess\PropertyPath;
 
 /**
  * @author Kévin Dunglas <dunglas@gmail.com>
+ *
+ * @final since Symfony 7.4
  */
 class AttributeMetadata implements AttributeMetadataInterface
 {
@@ -215,9 +217,9 @@ class AttributeMetadata implements AttributeMetadataInterface
     }
 
     /**
-     * Returns the names of the properties that should be serialized.
+     * @internal since Symfony 7.4, will be replaced by `__serialize()` in 8.0
      *
-     * @return string[]
+     * @final since Symfony 7.4, will be replaced by `__serialize()` in 8.0
      */
     public function __sleep(): array
     {