+++ Attribute/Context.php
@@ -19,7 +19,7 @@ use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 #[\Attribute(\Attribute::TARGET_CLASS | \Attribute::TARGET_PROPERTY | \Attribute::TARGET_METHOD | \Attribute::IS_REPEATABLE)]
 class Context
 {
-    private array $groups;
+    public readonly array $groups;
 
     /**
      * @param array<string, mixed> $context                The common context to use when serializing or deserializing
@@ -30,9 +30,9 @@ class Context
      * @throws InvalidArgumentException
      */
     public function __construct(
-        private readonly array $context = [],
-        private readonly array $normalizationContext = [],
-        private readonly array $denormalizationContext = [],
+        public readonly array $context = [],
+        public readonly array $normalizationContext = [],
+        public readonly array $denormalizationContext = [],
         string|array $groups = [],
     ) {
         if (!$context && !$normalizationContext && !$denormalizationContext) {
@@ -48,21 +48,25 @@ class Context
         }
     }
 
+    #[\Deprecated('Use the "context" property instead', 'symfony/serializer:7.4')]
     public function getContext(): array
     {
         return $this->context;
     }
 
+    #[\Deprecated('Use the "normalizationContext" property instead', 'symfony/serializer:7.4')]
     public function getNormalizationContext(): array
     {
         return $this->normalizationContext;
     }
 
+    #[\Deprecated('Use the "denormalizationContext" property instead', 'symfony/serializer:7.4')]
     public function getDenormalizationContext(): array
     {
         return $this->denormalizationContext;
     }
 
+    #[\Deprecated('Use the "groups" property instead', 'symfony/serializer:7.4')]
     public function getGroups(): array
     {
         return $this->groups;