+++ Mapping/Loader/AttributeLoader.php
@@ -44,14 +44,12 @@ class AttributeLoader implements LoaderInterface
     ];
 
     /**
-     * @param bool|null      $allowAnyClass Null is allowed for BC with Symfony <= 6
      * @param class-string[] $mappedClasses
      */
     public function __construct(
-        private ?bool $allowAnyClass = true,
+        private bool $allowAnyClass = true,
         private array $mappedClasses = [],
     ) {
-        $this->allowAnyClass ??= true;
     }
 
     /**