+++ Normalizer/BackedEnumNormalizer.php
@@ -77,7 +77,21 @@ final class BackedEnumNormalizer implements NormalizerInterface, DenormalizerInt
                 return null;
             }
 
-            throw NotNormalizableValueException::createForUnexpectedDataType('The data must belong to a backed enumeration of type '.$type, $data, ['int', 'string'], $context['deserialization_path'] ?? null, true, 0, $e);
+            $backingType = (new \ReflectionEnum($type))->getBackingType()->getName();
+
+            if ($e instanceof \TypeError || get_debug_type($data) !== $backingType) {
+                throw NotNormalizableValueException::createForUnexpectedDataType('The data must be of type '.$backingType, $data, [$backingType], $context['deserialization_path'] ?? null, true, 0, $e);
+            }
+
+            $expectedValues = array_map(function ($type) {
+                if (\is_string($type->value)) {
+                    return "'{$type->value}'";
+                }
+
+                return $type->value;
+            }, $type::cases());
+
+            throw new NotNormalizableValueException('The data must be one of the following values: '.implode(', ', $expectedValues), 0, $e, $type, null, $context['deserialization_path'] ?? null, true);
         }
     }