+++ CHANGELOG.md
@@ -23,7 +23,7 @@ CHANGELOG
    public function denormalize(string $propertyName, ?string $class = null, ?string $format = null, array $context = []): string;
    ```
  * Remove `AdvancedNameConverterInterface`, use `NameConverterInterface` instead
- * Remove the `CompiledClassMetadataFactory` and `CompiledClassMetadataCacheWarmer` classes
+ * Remove `ClassMetadataFactoryCompiler`, `CompiledClassMetadataFactory` and `CompiledClassMetadataCacheWarmer`
  * Remove class aliases in the `Annotation` namespace, use attributes instead
  * Remove getters in attribute classes in favor of public properties